requirements.dev.txt
anyio==4.12.1
asyncpg==0.31.0
click==8.3.1
databases==0.9.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
punq==0.7.0
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.45
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
